Implment first draft of force computation
Computation of the gravitational force on a pair of bodies is
imolemented. A class is provided for which the gravitational constant
can be set, which will be multiplied by the result of a generic
calculation of the structure of the gravitation force equation.

The gravitational constant in SI units has been specified in a separate
file, both as float and double.

Some refactoring of the vectors has been made to move generic
mathematical operations, such as abs, to a separate file.