Implement integration infterface
Several changes are made, most notably an interface for integrators is
implemented. A simple forward-euler integrator is added. A concrete
ForceComputer, doing a simple direct-sum computation on a single cpu
core, is also implemented.

It is clear that several refactors need to happen before moving forward.
The current interface does not supporth higher-order integrators because
these would need a new StarSystem, which would be a waste. In addition
the computation of accelerations is too complex because ForceComputer
has no access to the body masses.