Refactor force computation to accomodata potential
The force computation is partially refactored to allow for a
simultaneous computation of the gravitational potential when requested
by the user.

The current solution might be overengineered since at first there seems
to be essentially no runtime difference between computing the forces
alone and computing the forces and the potential energy.