Test for force computation and integration timer
A test is added for the computation of forces. It compares the result of
standalone force calculations and force-potential calculations.

A timing script for the ForwardEuler integrator is also added. This
revealed that when doing a large number of steps computing only forces
is notably more efficient than computing forces and potentials.