Implement second order Runge-Kutta integrator
The second order Runge-Kutta integrator is implemented. The integration
interface has to be slightly change not to be const anymoe since some
integrators will have to be stateful if they are to be efficient.

It has also been decided that the potential energy is always calculated
once during an integration step for now. During the intermediate stages
this is however not needed.